+++ src/calculator/EquationPanel.java
@@ -1,0 +1,25 @@
+package calculator;
+
+import javax.swing.JFrame;
+import javax.swing.JPanel;
+import javax.swing.JButton;
+import java.awt.event.ActionListener;
+import java.awt.event.ActionEvent;
+import javax.swing.JTextField;
+import javax.swing.JList;
+import javax.swing.JScrollBar;
+import javax.swing.JLabel;
+import javax.swing.JComboBox;
+import javax.swing.JFileChooser;
+
+public class EquationPanel extends JPanel {
+	
+	public EquationPanel() {
+		addKeys();
+	}
+	
+	public void addKeys() {
+		
+	}
+
+}
